[refactor] js 코드 분리
