+++ app/view/header.php
@@ -15,38 +15,4 @@
 <?php
 include_once 'app/view/modal.php';
 ?>
-<script>
-    const signinButton = document.getElementById('signin-button');
-    signinButton.addEventListener('click', () => submitSignin('app/view/signin.php'));
-
-    const signupButton = document.getElementById('signup-button');
-    signupButton.addEventListener('click', () => submitSignin('app/view/signup.php'));
-
-    function activeModal() {
-        const modal = document.getElementById('modal');
-        if (modal.style.visibility == 'visible')
-            modal.style.visibility = 'hidden'
-        else
-            modal.style.visibility = 'visible';
-    }
-
-    function submitSignin(path) {
-        const httpRequest = new XMLHttpRequest();
-
-        httpRequest.onreadystatechange = alertContents;
-        httpRequest.open('GET', path);
-        httpRequest.send();
-        activeModal();
-
-        function alertContents() {
-            if (httpRequest.readyState === XMLHttpRequest.DONE) {
-                if (httpRequest.status === 200) {
-                    const modalNode = document.getElementById('modal-content');
-                    modalNode.innerHTML = httpRequest.responseText;
-                } else {
-                    alert('request 실패');
-                }
-            }
-        }
-    }
-</script>
+<script src="app/view/header.js" type="module"></script>
